ProPS/requirements_props.txt
gymnasium==0.29.1
numpy>=1.24.0
matplotlib>=3.7.0
jinja2>=3.1.2
google-genai>=0.2.0
